\name{pasta}
\alias{pasta}
\docType{data}
\title{Pasta experiment}
\description{
    Dataset (R-Average form) of Pasta experiment.
    Factors: Price (3 levels: 0.89, 0.99, 1.09) x Packaging (3 levels).
}
\usage{data(pasta)}
\format{A \code{matrix} object.}

\examples{
# data(pasta)
}
\keyword{datasets}
