\name{pasta}
\alias{pasta}
\docType{data}
\title{Pasta experiment}
\description{
    Data of four subjects from the Pasta experiment. The table contains data of the one-way
    sub-designs, folowed by the data of the full-factorial design.
    Factors: Price (3 levels: 0.89, 0.99, 1.09) and Packaging (3 levels: box with window, box
    without window, plastic bag).
}
\usage{data(pasta)}
\format{A \code{data.frame} object.}

\references{
    Massidda D., Polezzi D., Vidotto G. (2011). A Functional Measurement approach to cope
    the non-linearity of judgments in marketing research. \emph{Proceedings of the 10th European
    Conference on Research Methodology for Business and Management Studies}. Normandy Business
    School, Caen, France, 348-354.
}

\examples{
data(pasta)
}
\keyword{datasets}
